nem sei

+++ Algo237ED/src/com/gpaganini/Num.java
@@ -2,5 +2,5 @@ package com.gpaganini;
 
 public class Num {
     int c, num;
-    float h;
+    double h;
 }
